[refactor] QueryDSL에서 위치 정보에 대한 비교를 더욱 수월하게 하기 위해 Point -> Double타입의 변수 두개로 리펙토링
