+++ src/main/java/com/capstone/liveAloneCommunity/domain/location/Location.java
@@ -1,14 +1,16 @@
-package com.capstone.liveAloneCommunity.dto.member;
+package com.capstone.liveAloneCommunity.domain.location;
 
+import jakarta.persistence.Embeddable;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
-@Getter
+@Embeddable
 @AllArgsConstructor
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class ChangeLocationRequestDto {
+@Getter
+public class Location {
 
     private Double latitude;
     private Double longitude;
