+++ src/main/java/com/capstone/liveAloneCommunity/entity/member/Member.java
@@ -1,11 +1,11 @@
 package com.capstone.liveAloneCommunity.entity.member;
 
+import com.capstone.liveAloneCommunity.domain.location.Location;
 import com.capstone.liveAloneCommunity.domain.member.*;
 import com.capstone.liveAloneCommunity.dto.auth.RegisterRequestDto;
 import jakarta.persistence.*;
 import lombok.*;
 import org.locationtech.jts.geom.Point;
-import org.locationtech.jts.io.WKTReader;
 import org.springframework.security.crypto.password.PasswordEncoder;
 
 @Entity
@@ -31,8 +31,8 @@ public class Member{
     private String providerId;
     @Column
     private String provider;
-    @Column
-    private Point point;
+    @Embedded
+    private Location location;
 
     @SneakyThrows
     public Member(RegisterRequestDto registerRequestDto, PasswordEncoder passwordEncoder){
@@ -41,7 +41,7 @@ public class Member{
         this.email = new Email(registerRequestDto.getEmail());
         this.password = new Password(passwordEncoder.encode(registerRequestDto.getPassword()));
         this.role = Role.USER;
-        this.point = (Point) new WKTReader().read(String.format("POINT(%f %f)", 37.22214637050458, 127.18653080961145));
+        this.location = new Location(37.22214637050458, 127.18653080961145);
     }
 
     @SneakyThrows
@@ -54,7 +54,7 @@ public class Member{
         this.role = role;
         this.provider = provider;
         this.providerId = providerId;
-        this.point = (Point) new WKTReader().read(String.format("POINT(%f %f)", 37.22214637050458, 127.18653080961145));
+        this.location = new Location(37.22214637050458, 127.18653080961145);
     }
 
     public String getUsername(){
@@ -85,11 +85,11 @@ public class Member{
         this.password = new Password(passwordEncoder.encode(newPassword));
     }
 
-    public Point getPoint(){
-        return point;
+    public Location getLocation(){
+        return location;
     }
 
-    public void changePoint(Point point) {
-        this.point = point;
+    public void changeLocation(Location location) {
+        this.location = location;
     }
 }
