+++ src/main/java/com/capstone/liveAloneCommunity/exception/member/LocationNotValidException.java
@@ -1,0 +1,4 @@
+package com.capstone.liveAloneCommunity.exception.member;
+
+public class LocationNotValidException extends RuntimeException {
+}
